+++ Client/Scripts/Gameplay/Dto/Snapshots/ProjectileSnapshot.cs
@@ -3,5 +3,5 @@ namespace Soteo.Gameplay.Dto.Snapshots;
 public sealed record ProjectileSnapshot : EntitySnapshot<ProjectileSnapshot>
 {
     public required DeflatedAbilityContext AbilityContext { get; init; }
-    public required float Speed { get; init; }
+    public required double Speed { get; init; }
 }
